fix: reset fusion params after HF load (fixes H100 loss=nan)
Root cause: from_pretrained(low_cpu_mem_usage=True, device_map=...) moves the
whole model to meta, discarding AerialVLAModel.__init__ init for the fresh
fusion stack, then materializes only checkpoint keys -> uninitialized memory
(BN weight ~8e35, LayerNorm NaN) -> NaN enters at lidar_encoder output on
CUDA bf16.

Add AerialVLAModel.reset_fusion_parameters() (kaiming conv/linear, BN reset,
LayerNorm=1) and call it right after get_peft_model in train_step_a.py.
Add regression tests (NaN recovery + bf16 preservation). H100 smoke now
passes with finite loss (10.77 -> 14.69).

+++ datasets/uav_lidar_dataset.py
@@ -710,9 +710,11 @@ class UAVLiDARCollator(object):
             "lidar_points": to_t([b["lidar_points"] for b in batch]),
             "lidar_valid": to_t([b["lidar_valid"] for b in batch]).bool(),
             # CEM wants K^-1 here (see module docstring / models/encoders/cem.py).
+            # linalg.inv has no bf16 kernel, and bf16 inversion loses precision,
+            # so invert in fp32 then cast back to the batch dtype.
             "camera_intrinsics": torch.linalg.inv(
-                to_t([b["camera_intrinsics"] for b in batch])
-            ),
+                to_t([b["camera_intrinsics"] for b in batch]).float()
+            ).to(dtype=self.dtype),
             "camera_extrinsics": to_t([b["camera_extrinsics"] for b in batch]),
             "normalised_pixel_coords": to_t([b["normalised_pixel_coords"] for b in batch]),
             "depth_samples": to_t([b["depth_samples"] for b in batch]),
